[GCS] Basic Integration of GCS
