+++ src/semantic_world/views/world_rdr/world_rdr.py
@@ -1,7 +1,7 @@
-from ...world import World
-from ripple_down_rules.datastructures.case import Case, create_case
 from typing_extensions import Any, Dict
 from ripple_down_rules.helpers import general_rdr_classify
+from ripple_down_rules.datastructures.case import Case, create_case
+from ...world import World
 from . import world_views_mcrdr as views_classifier
 
 classifiers_dict = dict()
