+++ test/test_worlds/test_gcs.py
@@ -1,16 +1,17 @@
 import os
-import time
 import unittest
 
-from matplotlib import pyplot as plt
+import plotly.graph_objects as go
 from random_events.interval import SimpleInterval
 from random_events.product_algebra import SimpleEvent
-import rustworkx as rx
 
-import plotly.graph_objects as go
-
-from semantic_world.graph_of_convex_sets import GraphOfConvexSets, BoundingBox, SpatialVariables, BoundingBoxCollection
+from semantic_world.adapters.urdf import URDFParser
+from semantic_world.graph_of_convex_sets import GraphOfConvexSets, SpatialVariables, PoseOccupiedError
+from semantic_world.geometry import BoundingBox, BoundingBoxCollection
+from semantic_world.prefixed_name import PrefixedName
 from semantic_world.spatial_types import Point3
+from semantic_world.world import World
+from semantic_world.world_entity import Body
 
 
 class GCSTestCase(unittest.TestCase):
@@ -29,12 +30,12 @@ class GCSTestCase(unittest.TestCase):
         z_lim = SimpleInterval(.45, .55)
         x_lim = SimpleInterval(-2, 3)
         y_lim = SimpleInterval(-2, 3)
-        limiting_event = SimpleEvent({SpatialVariables.z.value: z_lim,
-                                      SpatialVariables.x.value: x_lim,
-                                      SpatialVariables.y.value: y_lim})
+        limiting_event = SimpleEvent({SpatialVariables.x.value: x_lim,
+                                      SpatialVariables.y.value: y_lim,
+                                      SpatialVariables.z.value: z_lim,})
         obstacles = BoundingBoxCollection.from_event(
             ~obstacle.simple_event.as_composite_set() & limiting_event.as_composite_set())
-        [gcs.add(bb) for bb in obstacles]
+        [gcs.add_node(bb) for bb in obstacles]
         gcs.calculate_connectivity()
         cls.gcs = gcs
 
@@ -43,7 +44,7 @@ class GCSTestCase(unittest.TestCase):
         target_point = Point3.from_xyz(2, 2, 0.5)
 
         path = self.gcs.path_from_to(start_point, target_point)
-        self.assertEqual(len(path), 3)
+        self.assertEqual(len(path), 4)
 
     def test_plot(self):
         free_space_plot = go.Figure(self.gcs.plot_free_space())
@@ -57,30 +58,37 @@ class GCSFromWorldTestCase(unittest.TestCase):
     Test the application of a connectivity graph to the belief state.
     """
 
+    world: World
+
+    @classmethod
+    def setUpClass(cls):
+        urdf_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "..", "resources", "urdf")
+        apartment = os.path.join(urdf_dir, "table.urdf")
+        apartment_parser = URDFParser(apartment)
+        cls.world = apartment_parser.parse()
+
+
+
     def test_from_world(self):
         search_space = BoundingBox(min_x=-1, max_x=1,
                                    min_y=-1, max_y=1,
                                    min_z=0.1, max_z=1).as_collection()
         gcs = GraphOfConvexSets.free_space_from_world(self.world, search_space=search_space)
         self.assertIsNotNone(gcs)
-        self.assertGreater(len(gcs.nodes), 0)
-        self.assertGreater(len(gcs.edges), 0)
+        self.assertGreater(len(gcs.graph.nodes()), 0)
+        self.assertGreater(len(gcs.graph.edges()), 0)
 
-        start = PoseStamped.from_list([-0.9, -0.9, 0.4])
-        target = PoseStamped.from_list([-0.9, 0.9, 0.9])
+        start = Point3.from_xyz(-0.9, -0.9, 0.4)
+        target = Point3.from_xyz(-0.9, 0.9, 0.9)
 
         path = gcs.path_from_to(start, target)
-
-        if "ROS_VERSION" in os.environ:
-            pub = TrajectoryPublisher()
-            pub.visualize_trajectory(path)
 
         self.assertIsNotNone(path)
         self.assertGreater(len(path), 1)
 
         with self.assertRaises(PoseOccupiedError):
-            start = PoseStamped.from_list([-10, -10, -10])
-            target = PoseStamped.from_list([10, 10, 10])
+            start = Point3.from_xyz(-10, -10, -10)
+            target = Point3.from_xyz(10, 10, 10)
             gcs.path_from_to(start, target)
 
     def test_navigation_map_from_world(self):
@@ -88,8 +96,8 @@ class GCSFromWorldTestCase(unittest.TestCase):
                                    min_y=-1, max_y=1,
                                    min_z=0.1, max_z=1)
         gcs = GraphOfConvexSets.navigation_map_from_world(self.world, search_space=BoundingBoxCollection([search_space]))
-        self.assertGreater(len(gcs.nodes), 0)
-        self.assertGreater(len(gcs.edges), 0)
+        self.assertGreater(len(gcs.graph.nodes()), 0)
+        self.assertGreater(len(gcs.graph.edges()), 0)
 
 
 if __name__ == '__main__':
